+++ .travis.yml
@@ -1,14 +1,14 @@
-language: node_js
-sudo: false
-node_js:
-  - "7"
-script:
-- yarn run cover
-deploy:
-  provider: npm
-  email: "tparnell8@gmail.com"
-  skip_cleanup: true
-  api_key:
-    secure: teFFjO6l0OJVY53aGiV8ri51Ke3YEyQ3a7+vlYdgpoGWOecQl2wd8K9rRJZV19BKzGn/m+dMlydYYmS8v4lRIC0mMAv9g2cga65VMcR15gU/dFH71ujAxNDwx/JCey1HJooZnLX5kju8r3fI/i8mHoXaskUEJ5Kj9JXAPpDfkOflGaAZ2yR663geUf82iqH01zwAdFLtDh5jVmN1IeqYLB74J3PYpm65hH8NzN/Ljkk67mX8TnPIssH+PMLI2NBZqxR7MV8AxT/G9m28VZVXvjUTLbW0McpfE4YgJy1axT1OLTHOW/5XdHkPRLIUBARrzVrwrREZhYE27jRgcT9Dfr57yddXuz9kSZN4h1WB9bWbiX3CYgSnFoxegDGryeRJuL30lyzZWzNSB41pm0yhGM1xQGUznc8RRcQvSpjWeb/K+bkBxM65m9PzkX3b6CqsxTXNqNFIBNug6oID+KB9yNPOzpnvpGMhp29/cErTidTmcDSKQD6buFBAcPrRlGj44MRKjdpvNJGcUw0rkb9SgAXJnYYpK2/UTxhiY4+s6SmP0XMGuxsbng5Zk+zRNl53E0p2k1a9h4LXt+UCjwKtr6u5RI7Fqq6yZ7cj2xk6M6UlZzZmg4lVnVdbCODFI9zCWQw/Va+QlwtnA5PWhEyPA/RYiniiUoxftc/moH/8Rcs=
-  on:
-    tags: true
+# language: node_js
+# sudo: false
+# node_js:
+#   - "7"
+# script:
+# - yarn run cover
+# deploy:
+#   provider: npm
+#   email: "tparnell8@gmail.com"
+#   skip_cleanup: true
+#   api_key:
+#     secure: teFFjO6l0OJVY53aGiV8ri51Ke3YEyQ3a7+vlYdgpoGWOecQl2wd8K9rRJZV19BKzGn/m+dMlydYYmS8v4lRIC0mMAv9g2cga65VMcR15gU/dFH71ujAxNDwx/JCey1HJooZnLX5kju8r3fI/i8mHoXaskUEJ5Kj9JXAPpDfkOflGaAZ2yR663geUf82iqH01zwAdFLtDh5jVmN1IeqYLB74J3PYpm65hH8NzN/Ljkk67mX8TnPIssH+PMLI2NBZqxR7MV8AxT/G9m28VZVXvjUTLbW0McpfE4YgJy1axT1OLTHOW/5XdHkPRLIUBARrzVrwrREZhYE27jRgcT9Dfr57yddXuz9kSZN4h1WB9bWbiX3CYgSnFoxegDGryeRJuL30lyzZWzNSB41pm0yhGM1xQGUznc8RRcQvSpjWeb/K+bkBxM65m9PzkX3b6CqsxTXNqNFIBNug6oID+KB9yNPOzpnvpGMhp29/cErTidTmcDSKQD6buFBAcPrRlGj44MRKjdpvNJGcUw0rkb9SgAXJnYYpK2/UTxhiY4+s6SmP0XMGuxsbng5Zk+zRNl53E0p2k1a9h4LXt+UCjwKtr6u5RI7Fqq6yZ7cj2xk6M6UlZzZmg4lVnVdbCODFI9zCWQw/Va+QlwtnA5PWhEyPA/RYiniiUoxftc/moH/8Rcs=
+#   on:
+#     tags: true
