+++ test/package.json
@@ -10,6 +10,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {    
-    "gulp": "^3.9.1"
+    "gulp": "^4.0.2"
   }
 }
