+++ test/tst/UnitTest1.cs
@@ -1,14 +1,14 @@
 using System;
 using Xunit;
 
-namespace tstproj
+namespace tst
 {
     public class UnitTest1
     {
         [Fact]
         public void Test1()
         {
-
+            Assert.Equal(1, 1);
         }
     }
 }
